[ADD] Shm

+++ src/main.c
@@ -45,7 +45,7 @@ void close_child(int signal)
 */
 int		main(int argc, char **argv)
 {
-	t_data_info *data;
+	// t_data_info *data;
 	time_t now;
 	struct tm tm_now;
 	char *data_file_path;// = "../Space_MIDI/data_files";
@@ -63,11 +63,11 @@ int		main(int argc, char **argv)
 	}
 	else
 	{
-		data_file_path = (char*)malloc(sizeof(char) * (sizeof("../Space_MIDI/data_files/") + 1));
-		data_file_path = strcpy(data_file_path, "../Space_MIDI/data_files/");
+		data_file_path = (char*)malloc(sizeof(char) * (sizeof("./Space_MIDI/data_files/") + 1));
+		data_file_path = strcpy(data_file_path, "./Space_MIDI/data_files/");
 		
-		midi_file_path = (char*)malloc(sizeof(char) * (sizeof("../Space_MIDI/midi_files/") + 1));
-		midi_file_path = strcpy(midi_file_path, "../Space_MIDI/midi_files/");
+		midi_file_path = (char*)malloc(sizeof(char) * (sizeof("./Space_MIDI/midi_files/") + 1));
+		midi_file_path = strcpy(midi_file_path, "./Space_MIDI/midi_files/");
 	}
 
 
@@ -83,7 +83,7 @@ int		main(int argc, char **argv)
 		  *  "../Space_MIDI/data_files", "../Space_MIDI/midi_files", NULL);
 		*/
 	
-		execl("../Space_MIDI/midi_controller", "midi_controller",\
+		execl("./Space_MIDI/midi_controller", "midi_controller",\
 		 data_file_path, midi_file_path, NULL);
 
 		exit(1);
@@ -92,16 +92,43 @@ int		main(int argc, char **argv)
 	// kill(pid, SIGKILL);SIGSTOP
 	signal(SIGINT, (void (*)(int))close_child);
 
+
+	t_circular_buffer *buffer;
+
 	//Memory Allocation
-	data = (t_data_info *)malloc(sizeof(t_data_info) * SAMPLES_NU);
-	for (int i = 0; i < SAMPLES_NU; i++)
-	{
-		data[i].time = (char*)malloc(sizeof("AAAA/mm/JJ HH:MM:SS"));
-		data[i].datas = (t_data *)malloc(sizeof(t_data) * DATAS_SIZE);
-	}
+	// data = (t_data_info *)malloc(sizeof(t_data_info) * SAMPLES_NU);
+	// for (int i = 0; i < SAMPLES_NU; i++)
+	// {
+	// 	data[i].time = (char*)malloc(sizeof("AAAA/mm/JJ HH:MM:SS"));
+	// 	data[i].datas = (t_data *)malloc(sizeof(t_data) * DATAS_SIZE);
+	// }
+
+   int shmid, numtimes;
+   struct shmseg *shmp;
+   char *bufptr;
+//    int spaceavailable;
+   shmid = shmget(SHM_KEY, 3000, 0644 | IPC_CREAT);
+
+   if (shmid == -1)
+   {
+
+      printf("Error code :%d \n", errno);
+      perror("Shared memory");
+      return 1;
+   }
+
+   shmp = shmat(shmid, NULL, 0);
+   if (shmp == (void *)-1)
+   {
+      perror("Shared memory attach");
+      return 1;
+   }
+
+   bufptr = shmp->buf;
 
 
-	int max_steps = FILES_NU * SAMPLES_NU; // 20
+
+	int max_steps = SAMPLES_NU; // 20
 	int current_step = 0;
 
 	int number_of_data_steps = (sizeof(g_all_data) / sizeof(t_data*)) - 1 ;
@@ -109,15 +136,20 @@ int		main(int argc, char **argv)
 	int current_data_step_min = 0;
 	int current_data_step_max = 0;
 
-
+	int buffer_index = 0;
 
 	//Writing Eatch File
-	for (int current_file = 0; current_file < FILES_NU; current_file++)
-	{
+	// for (int current_file = 0; current_file < FILES_NU; current_file++)
+	// {
 
 		//Data Feeding
 		for (int i = 0; i < SAMPLES_NU; i++, current_step++)
 		{
+			shmp->cnt = sizeof(t_circular_buffer);
+			buffer = (t_circular_buffer*)shmp->buf;
+			bzero(buffer, sizeof(t_circular_buffer));
+
+			buffer_index = i % 10;
 			int n = 1;
 			while (!(((max_steps / number_of_data_steps) * (n - 1) <= current_step)
 			&& ((max_steps / number_of_data_steps) * (n) > current_step)))
@@ -136,22 +168,25 @@ int		main(int argc, char **argv)
 			printf("N : %d, duty : %f\n",n, duty);
 			printf("step min : %d\n",current_data_step_min);
 			printf("step max : %d\n",current_data_step_max);
-			printf("current step : %d/%d\n\n", current_step, SAMPLES_NU * FILES_NU);
+			printf("current step : %d/%d\n\n", current_step, SAMPLES_NU);
 
 
 			for (int u = 0; u < DATAS_SIZE; u++)
 			{
 				// data[i].datas[u] = g_datas[u];
-				data[i].datas[u] = g_all_data[n - 1][u];
+				// data[i].datas[u] = g_all_data[n - 1][u];
 
 
-				if (data[i].datas[u].data_type == FLOATING)
+				if (g_all_data[n - 1][u].data_type == FLOATING)
 				{
-					data[i].datas[u].float_data = g_all_data[n - 1][u].float_data * (1-duty) + g_all_data[n][u].float_data * duty;
+					float *ptr = (float*)((uint64_t)&buffer->data[buffer_index] + g_all_data[n - 1][u].offset);
+
+
+					*ptr = g_all_data[n - 1][u].float_data * (1-duty) + g_all_data[n][u].float_data * duty;
 					float tmpdelta = g_all_data[n - 1][u].float_delta * (1-duty) + g_all_data[n][u].float_delta * duty;
 					if (tmpdelta > 0)
 					{
-						data[i].datas[u].float_data += fmod((float)rand() / 13., tmpdelta * 2) - (tmpdelta);
+						*ptr += fmod((float)rand() / 13., tmpdelta * 2) - (tmpdelta);
 					}
 					// if (u == 0)
 					// {
@@ -159,54 +194,79 @@ int		main(int argc, char **argv)
 					// 	// printf("tmpdelta : ")
 					// }
 				}
-				else if (data[i].datas[u].data_type == INTEGER)
+				else if (g_all_data[n - 1][u].data_type == INTEGER)
 				{
-					data[i].datas[u].int_data = (float)g_all_data[n - 1][u].int_data * (1-duty) + g_all_data[n][u].int_data * duty;
+					uint32_t *ptr = (uint32_t*)((uint64_t)&buffer->data[buffer_index] + g_all_data[n - 1][u].offset);
+
+					*ptr = (float)g_all_data[n - 1][u].int_data * (1-duty) + g_all_data[n][u].int_data * duty;
 					int tmpdelta = (float)g_all_data[n - 1][u].int_delta * (1-duty) + g_all_data[n][u].int_delta * duty;
 					if (tmpdelta > 0)
 					{
-						data[i].datas[u].int_data += (rand() % (tmpdelta * 2)) - tmpdelta;
+						*ptr += (rand() % (tmpdelta * 2)) - tmpdelta;
 					}
 				}
-				else if (data[i].datas[u].data_type == BINARY)
+				else if (g_all_data[n - 1][u].data_type == BINARY)
 				{
-					// data[i].datas[u].binary += rand() % 2;
+					uint8_t *ptr = (uint8_t*)((uint64_t)&buffer->data[buffer_index] + g_all_data[n - 1][u].offset);
+					*ptr = g_all_data[n][u].binary;
 				}
 			}
 			now = time(NULL);
-			tm_now = *localtime(&now);
+			buffer->data[buffer_index].timestamp = time(NULL);
+			printf("ORIGINAL TIMESTAMP : %llu\n", buffer->data[buffer_index].timestamp);
+		      shmp->complete = 0;
+		      bufptr = shmp->buf;
+
+			// tm_now = *localtime(&buffer->data[buffer_index].timestamp);//
 			// char s_now[sizeof("AAAA/mm/JJ HH:MM:SS")];
-			strftime(data[i].time, sizeof("AAAA/mm/JJ HH:MM:SS"), "%Y/%m/%d %H:%M:%S", &tm_now);
+			
+			// strftime(data[i].time, sizeof("AAAA/mm/JJ HH:MM:SS"), "%Y/%m/%d %H:%M:%S", &tm_now);//
 			// data[i].time = strdup(s_now);
 
 			// printf("Time : %s\n", s_now);
 			printf("Sample n°%d/%d acquired\n", i + 1, SAMPLES_NU);
 			sleep(SAMPLE_RATE);
 		}
+   shmp->complete = 1;
+
+
+
+      if (shmdt(shmp) == -1)
+   {
+      perror("shmdt");
+      return 1;
+   }
+
+   if (shmctl(shmid, IPC_RMID, 0) == -1)
+   {
+      perror("shmctl");
+      return 1;
+   }
+   printf("Writing Process: Complete\n");
 
 		//Files Saving
-		char s_filename[sizeof("AAAA_mm_JJ__HH_MM_SS.json")];
-		strftime(s_filename, sizeof(s_filename), "%Y_%m_%d__%H_%M_%S.json", &tm_now);
-		// if (argc > 1)
-		// {
-			char s_filepath[sizeof(s_filename) + strlen(data_file_path)];
-			sprintf(s_filepath, "%s%s", data_file_path, s_filename);
-			write_json(data, s_filepath);
-		// }
-		// else
-		// {
-		// 	write_json(data, s_filename);
-		// }
-		printf("File n°%d/%d writed\n", current_file + 1, FILES_NU);
-	}
+	// 	char s_filename[sizeof("AAAA_mm_JJ__HH_MM_SS.json")];
+	// 	strftime(s_filename, sizeof(s_filename), "%Y_%m_%d__%H_%M_%S.json", &tm_now);
+	// 	// if (argc > 1)
+	// 	// {
+	// 		char s_filepath[sizeof(s_filename) + strlen(data_file_path)];
+	// 		sprintf(s_filepath, "%s%s", data_file_path, s_filename);
+	// 		write_json(data, s_filepath);
+	// 	// }
+	// 	// else
+	// 	// {
+	// 	// 	write_json(data, s_filename);
+	// 	// }
+	// 	printf("File n°%d/%d writed\n", current_file + 1, FILES_NU);
+	// }
 
 	//Freeing Memory
-	for (int i = 0; i < SAMPLES_NU; i++)
-	{
-		free(data[i].time);
-		free(data[i].datas);
-	}
-	free(data);
+	// for (int i = 0; i < SAMPLES_NU; i++)
+	// {
+	// 	free(data[i].time);
+	// 	free(data[i].datas);
+	// }
+	// free(data);
 	sleep(1);
 	kill(g_pid, SIGTERM);
 	sleep(1);
